refacto enum et tests
