+++ domain/enum/NotificationEnum.ts
@@ -1,0 +1,10 @@
+export enum NotificationType {
+    INFO = "info",
+    ALERTE = "alerte",
+    ERREUR = "erreur",
+}
+
+export enum NotificationStatus {
+    READ = "read",
+    UNREAD = "unread",
+}
